move loading product to cartItem
